refactor(llmq): break circular dependencies in quorumproofs
Move QuorumMerkleProof and QuorumProofData structs to a new header
file (quorumproofdata.h) to break circular dependencies between
llmq/blockprocessor, llmq/quorumproofs, evo/cbtx, and llmq/quorumsman.

Also fixes trailing whitespace in quorumproofs.cpp.
